ref(index) : move data related to github in module
